modified components: planet-details and starship-details
