+++ src/components/sw-components/index.js
@@ -1,4 +1,6 @@
 import {PeopleList, PlanetList, StarshipList} from './item-lists';
-import PersonDetails from './person-details'
+import PersonDetails from './person-details';
+import PlanetDetails from './planet-details';
+import StarshipDetails from "./starship-details";
 
-export {PeopleList, PlanetList, StarshipList, PersonDetails}
+export {PeopleList, PlanetList, StarshipList, PersonDetails, PlanetDetails, StarshipDetails}
